Ret-3814 continously improvement, resolving PDF Mapper sonar lint problems  (#393)
* RET-3506 (#370)

* RET-3476 Elastic search max limit update (#334)

* Updated elastic search max size to 10000

* Updated elastic search max size to 10000

* RET-3476 Resolving Vulnerabilities (#337)

* Updated elastic search max size to 10000

* Updated elastic search max size to 10000

* Updated elastic search max size to 10000

* RET-3476 Functional Test Update (#338)

* Updated elastic search max size to 10000

* Updated elastic search max size to 10000

* Updated elastic search max size to 10000

* Functional test update. A user case list should return only one case when any case_id is given.

* Functional test update. A user case list should return at least one case when any case_id is given.

* RET-3098 - Java and app insights update (#333)

* RET-3098

* RET-3098

* RET-3098

* RET-3098

* RET-3098

* RET-3098

* RET-3098

* RET-3098

* RET-3098

---------


* Ret 3098 (#339)

* RET-3495 Adding timeout to functional test (#345)

* Add timeout to functional test stage3GetAllCaseDetailsShouldReturnAllCaseDetails

* Bumping chart version/ fixing aliases

* Implemented line 81 as SECONDS.sleep, not to have LawOfDemeter problem on pmd check.

* Implemented line 81 as SECONDS.sleep, not to have LawOfDemeter problem on pmd check.

---------


* Merge master into pet (#349)

* Update all dependencies (#323)

* Update all dependencies

* Bumping chart version/ fixing aliases

* Update gradle-wrapper.properties

* Bumping chart version/ fixing aliases

* Fix breaking versions

* Bump commons-fileupload

* Remove suppression

---------


* ES Fix (#341)

* Add timeout to functional test (#344)

* Add timeout to functional test

* Remove body

* CVE Fixes (#348)

* Checkstyle

---------


* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* RET-3613 (#368)

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Updates for PDFMapperUtil and Code beautify.

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

* Merged Updates

---------


* RET-3702 (#384)

* CVE fix

* CVE fix

* CVE fix

* CVE fix

---------


* build fix (#389)


* Update build.gradle (#392)

* Update build.gradle

* Update build.gradle

* Update suppressions.xml

* Changed packages of the exception(PdfServiceException) and model(PdfDecodedMultipartFile) classes

* Changed name of the class PdfMapperEmploymentUtil to PdfMapperEmploymentDetailsUtil

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

* Sonar Lint Updates

---------
