+++ src/main/java/uk/gov/hmcts/reform/et/syaapi/service/utils/PdfMapperEmploymentDetailsUtil.java
@@ -19,9 +19,9 @@ import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.WEE
 import static uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants.YES;
 
 
-public final class PdfMapperEmploymentUtil {
+public final class PdfMapperEmploymentDetailsUtil {
 
-    private PdfMapperEmploymentUtil() {
+    private PdfMapperEmploymentDetailsUtil() {
         // Utility classes should not have a public or default constructor.
     }
 
@@ -39,7 +39,7 @@ public final class PdfMapperEmploymentUtil {
                 GenericServiceUtil.logException("An error occurred while printing employment details to pdf file",
                                                 caseData.getEthosCaseReference(),
                                                 e.getMessage(),
-                                                "PdfMapperEmploymentUtil",
+                                                "PdfMapperEmploymentDetailsUtil",
                                                 "putEmploymentDetails");
             }
         }
