+++ src/test/java/uk/gov/hmcts/reform/et/syaapi/service/pdf/util/PdfMapperRespondentUtilTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.et.syaapi.service.utils;
+package uk.gov.hmcts.reform.et.syaapi.service.pdf.util;
 
 import org.apache.commons.lang3.StringUtils;
 import org.elasticsearch.common.Strings;
@@ -14,9 +14,16 @@ import uk.gov.hmcts.et.common.model.ccd.Address;
 import uk.gov.hmcts.et.common.model.ccd.CaseData;
 import uk.gov.hmcts.et.common.model.ccd.items.RespondentSumTypeItem;
 import uk.gov.hmcts.et.common.model.ccd.types.RespondentSumType;
-import uk.gov.hmcts.reform.et.syaapi.model.TestData;
 import uk.gov.hmcts.reform.et.syaapi.models.AcasCertificatePdfFieldModel;
 import uk.gov.hmcts.reform.et.syaapi.models.RespondentPdfFieldModel;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.GenericServiceUtil;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperConstants;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperRespondentUtil;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.PdfMapperServiceUtil;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.PdfTemplateRespondentFieldNamesEnum;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.data.PdfMapperRespondentUtilTestDataProvider;
+import uk.gov.hmcts.reform.et.syaapi.service.utils.data.TestDataProvider;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,22 +34,7 @@ import java.util.concurrent.ConcurrentMap;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.times;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.ADDRESS_LINE_1;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.ADDRESS_LINE_2;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.ADDRESS_LINE_3;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.COUNTRY;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.COUNTY;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.INTEGER_NUMERIC_FIVE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.INTEGER_NUMERIC_FOUR;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.INTEGER_NUMERIC_THREE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.INTEGER_NUMERIC_TWO;
 import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.NO;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.NULL_ADDRESS;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.NULL_STRING;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.POSTCODE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.POST_TOWN;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.STRING_NUMERIC_ONE;
-import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.TEST_COMPANY_NAME;
 import static uk.gov.hmcts.reform.et.syaapi.service.utils.TestConstants.YES;
 
 class PdfMapperRespondentUtilTest {
@@ -50,7 +42,8 @@ class PdfMapperRespondentUtilTest {
     @ParameterizedTest
     @NullSource
     @MethodSource(
-        "uk.gov.hmcts.reform.et.syaapi.model.TestData#generateCaseDataSamplesWithRespondentSumTypeItems")
+        "uk.gov.hmcts.reform.et.syaapi.service.utils.data.PdfMapperRespondentUtilTestDataProvider#"
+            + "generateCaseDataSamplesWithRespondentSumTypeItems")
     void putRespondent(CaseData respondentCaseData) {
         ConcurrentMap<String, Optional<String>> printFields = new ConcurrentHashMap<>();
         PdfMapperRespondentUtil.putRespondents(respondentCaseData, printFields);
@@ -66,7 +59,7 @@ class PdfMapperRespondentUtilTest {
             checkAcasCertificate(printFields,
                                  firstRespondent,
                                  firstRespondentPdfFieldModel.respondentAcasCertificatePdfFieldModel());
-            if (respondentCaseData.getRespondentCollection().size() >= INTEGER_NUMERIC_TWO) {
+            if (respondentCaseData.getRespondentCollection().size() >= TestConstants.INTEGER_NUMERIC_TWO) {
                 RespondentPdfFieldModel secondRespondentPdfFieldModel =
                     PdfTemplateRespondentFieldNamesEnum.SECOND_RESPONDENT.respondentPdfFieldModel;
                 RespondentSumType secondRespondent = respondentCaseData.getRespondentCollection().get(1).getValue();
@@ -75,7 +68,7 @@ class PdfMapperRespondentUtilTest {
                                      secondRespondent,
                                      secondRespondentPdfFieldModel.respondentAcasCertificatePdfFieldModel());
             }
-            if (respondentCaseData.getRespondentCollection().size() >= INTEGER_NUMERIC_THREE) {
+            if (respondentCaseData.getRespondentCollection().size() >= TestConstants.INTEGER_NUMERIC_THREE) {
                 RespondentPdfFieldModel thirdRespondentPdfFieldModel =
                     PdfTemplateRespondentFieldNamesEnum.THIRD_RESPONDENT.respondentPdfFieldModel;
                 RespondentSumType thirdRespondent = respondentCaseData.getRespondentCollection().get(2).getValue();
@@ -84,7 +77,7 @@ class PdfMapperRespondentUtilTest {
                                      thirdRespondent,
                                      thirdRespondentPdfFieldModel.respondentAcasCertificatePdfFieldModel());
             }
-            if (respondentCaseData.getRespondentCollection().size() >= INTEGER_NUMERIC_FOUR) {
+            if (respondentCaseData.getRespondentCollection().size() >= TestConstants.INTEGER_NUMERIC_FOUR) {
                 RespondentPdfFieldModel forthRespondentPdfFieldModel =
                     PdfTemplateRespondentFieldNamesEnum.FORTH_RESPONDENT.respondentPdfFieldModel;
                 RespondentSumType forthRespondent = respondentCaseData.getRespondentCollection().get(3).getValue();
@@ -93,7 +86,7 @@ class PdfMapperRespondentUtilTest {
                                      forthRespondent,
                                      forthRespondentPdfFieldModel.respondentAcasCertificatePdfFieldModel());
             }
-            if (respondentCaseData.getRespondentCollection().size() >= INTEGER_NUMERIC_FIVE) {
+            if (respondentCaseData.getRespondentCollection().size() >= TestConstants.INTEGER_NUMERIC_FIVE) {
                 RespondentPdfFieldModel fifthRespondentPdfFieldModel =
                     PdfTemplateRespondentFieldNamesEnum.FIFTH_RESPONDENT.respondentPdfFieldModel;
                 RespondentSumType fifthRespondent = respondentCaseData.getRespondentCollection().get(4).getValue();
@@ -108,19 +101,24 @@ class PdfMapperRespondentUtilTest {
     @Test
     void putRespondentLogsPdfServiceExceptionWhenWrongNoAcasReasonSelected() {
         try (MockedStatic<GenericServiceUtil> mockedServiceUtil = Mockito.mockStatic(GenericServiceUtil.class)) {
-            Address respondentAddress = TestDataProvider.generateAddressByAddressFields(ADDRESS_LINE_1,
-                                                                                        ADDRESS_LINE_2,
-                                                                                        ADDRESS_LINE_3,
-                                                                                        POST_TOWN, COUNTY,
-                                                                                        COUNTRY, POSTCODE);
+            Address respondentAddress = TestDataProvider.generateAddressByAddressFields(TestConstants.ADDRESS_LINE_1,
+                                                                                        TestConstants.ADDRESS_LINE_2,
+                                                                                        TestConstants.ADDRESS_LINE_3,
+                                                                                        TestConstants.POST_TOWN,
+                                                                                        TestConstants.COUNTY,
+                                                                                        TestConstants.COUNTRY,
+                                                                                        TestConstants.POSTCODE);
             CaseData respondentCaseData =
-                TestDataProvider.generateCaseDataForRespondent(STRING_NUMERIC_ONE, YES, NULL_ADDRESS);
+                TestDataProvider.generateCaseDataForRespondent(TestConstants.STRING_NUMERIC_ONE,
+                                                               YES,
+                                                               TestConstants.NULL_ADDRESS);
             RespondentSumTypeItem respondentSumTypeItem =
-                TestData.generateRespondentSumTypeItem(STRING_NUMERIC_ONE, TEST_COMPANY_NAME,
-                                                       respondentAddress,
-                                                       NO,
-                                                       NULL_STRING,
-                                                       "DUMMY REASON");
+                PdfMapperRespondentUtilTestDataProvider.generateRespondentSumTypeItem(TestConstants.STRING_NUMERIC_ONE,
+                                                                                      TestConstants.TEST_COMPANY_NAME,
+                                                                                      respondentAddress,
+                                                                                      NO,
+                                                                                      TestConstants.NULL_STRING,
+                                                                                      "DUMMY REASON");
 
             List<RespondentSumTypeItem> respondentCollection = new ArrayList<>();
             respondentCollection.add(respondentSumTypeItem);
